#13 game results window

See merge request lior-gleizer/trivia_lior_itay!81

+++ Trivia/SqliteDatabase.h
@@ -68,5 +68,5 @@ private:
 struct ScoreData {
 	int correctAnswers;
 	int totalAnswers;
-	int averageTime;
+	double averageTime;
 };
